Move getRemediationURLFromMetadata to common lib
